Separate analytics and sockets into private comps
Socket is now a private variable owned solely by the Ting component.
Children can communicate user intentions to the parent for network
propagation through parent callbacks. The parent can inform the children
of state changes due to network actions through event firing. This
allows us to encapsulate the web-socket network specifics and protocol
specification into one file and separate concerns clearly.

We also modify analytics code in this commit. Analytics is now
encapsulated in the analytics file and user intentions are communicated
with event callbacks.

These two features are merged in one commit because the architectural
changes of using event firing to designate user intentions is similar
for both and the code used is parallel.